projects: restore support for stars & update sort
